wasm: add an Emscripten build target for fc, sfc, and md
Cross-compiles the fc, sfc, and md cores to WebAssembly behind a small C ABI
that loads a ROM, runs one frame at a time, and exposes the video, audio, and
error buffers. Headless: the frontend lives in JavaScript.

The port needs four things from the existing tree.

libco gains an Emscripten backend built on emscripten_fiber_swap, selected in
libco.c on __EMSCRIPTEN__. Fiber stacks grow to 128 KiB because Asyncify stores
its unwind state on the coroutine stack.

nall learns wasm32 as an architecture and web as a platform, which introduces
the PLATFORM_WEB macro and an ARCHITECTURE_WASM32 struct alongside the existing
detections. wasm is little-endian, so it joins that branch of the endian test.

The SH2 recompiler is guarded behind SLJIT, which has no wasm backend. A stub
Recompiler with `enabled = false` keeps the interpreter path compiling
unchanged; every call site is already behind a `recompiler.enabled` or
`Accuracy::Recompiler` test.

Two 32-bit-target fixes: mia's braced initializers narrowed u64 sizes into a
32-bit size_t, which is ill-formed in a braced-init-list, and nall/decode/mmi.hpp
was missing a `#pragma once`.

Native builds are unaffected. Every core change is behind __EMSCRIPTEN__ or
SLJIT, and the 32-bit fixes are no-ops where size_t is already 64-bit.

+++ mia/pak/pak.cpp
@@ -118,7 +118,7 @@ auto Pak::save(string name, string extension, string location) -> bool {
     auto saveFilePath = saveLocation(location, name, extension);
     directory::create(Location::dir(saveFilePath));
     if(auto fp = file::open(saveLocation(location, name, extension), file::mode::write)) {
-      fp.write({save->data(), save->size()});
+      fp.write({save->data(), (size_t)save->size()});
       return true;
     }
   }
